video_core: Move MortonCopyPixels128 to utils header.
